Add locking for block group scans list. Still TODO is the list of buffers in the block group.
Made a few other improvements: block group scan list now points to the scan hash entry as well for fast lookups. (can probably remove the scan ID, and the entire map... but do this later, once we're sure it won't be needed for something else. Should be low-overhead for now)